requirements.txt
# Gemini (voice assistant - fast responses)
google-genai>=1.0.0

# Web server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0
jinja2>=3.1.0

# Smart home (Tuya/SmartLife)
tuya-connector-python>=0.1.2

# Desktop agent (only needed on MacBook)
psutil>=5.9.0
